Created PostsPage and moved compoent content to PostsPage from App.
