+++ instagram/src/App.js
@@ -1,44 +1,18 @@
 import React, { Component } from 'react';
 import './App.css';
-import dummyData from './dummy-data';
-import PostContainer from './components/PostContainer/PostContainer';
-import SearchBar from './components/SearchBar/SearchBar';
+
+import PostsPage from './components/PostContainer/PostsPage'
 
 class App extends Component {
   constructor() {
-    super();
-    this.state = {
-      posts: [],
-      filtered: []
-    };
+    super()
+    this.state = {}
   }
-  componentDidMount() {
-    this.setState({ posts: dummyData });
-  }
-  searchPostsHandler = event => {
-    const posts = this.state.posts.filter(post => {
-      if (post.username.includes(event.target.value)) {
-        return post;
-      }
-    });
-    this.setState({ filtered: posts });
-  };
+
   render() {
     return (
-      <div className="App">
-        <SearchBar
-          searchTerm={this.state.searchTerm}
-          searchPosts={this.searchPostsHandler}
-        />
-        <PostContainer
-          posts={
-            this.state.filtered.length > 0
-              ? this.state.filtered
-              : this.state.posts
-          }
-        />
-      </div>
-    );
+      <PostsPage />
+    )
   }
 }
